Improve logging of validation errors
The check_validate step in compileProgram was throwing
RuntimeError: Validation failure (0), with no description
of the error.

Disabling it gives more details in the RuntimeError. See
https://sourceforge.net/p/pyopengl/bugs/258/

This option isn't available in the version of PyOpenGL I
have installed, so I've just copied it in from the latest
version for now. See
https://github.com/mcfletch/pyopengl/blob/0d71fcf307cc9c5f7a856d736ff6f502da1e8bc0/OpenGL/GL/shaders.py#L205