Fixes and enhancements.
NEW Maya look and feel.
NEW Zoom in/out with the mouse wheel.
FIX Y-axis is inverted (towards +infinity is downwards).
FIX To operate with a key, two independent clicks are needed (one to select it, another one to drag it).
FIX Selection marquee is not displayed.
FIX Curves do not cross the key origin.